+++ next.config.ts
@@ -2,6 +2,12 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   output: "standalone",
+  outputFileTracingIncludes: {
+    "/*": [
+      "node_modules/@prisma/client/**/*",
+      "node_modules/.prisma/client/**/*",
+    ],
+  },
   outputFileTracingExcludes: {
     "*": [
       ".env*",